packages/etsy-core/pyproject.toml
[project]
name = "etsy-core"
version = "0.0.0"
description = "Low-level Etsy API connectivity: OAuth 2.0 + PKCE, httpx async client, rate limiter, retry, F3 redaction. No MCP dependency."
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.0",
    "pydantic>=2.0.0",
    "tenacity>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/etsy_core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
]
